+++ examples/acquired_immunity/acquired_immunity.jl
@@ -2,7 +2,7 @@
 # julia --project=. examples/acquired_immunity/acquired_immunity.jl
 # unless viewing flamegraph, then run from console
 
-using Revise
+# using Revise
 using jOpqua
 
 using StaticArrays
@@ -18,22 +18,21 @@ function run(seed::Int64, t_vec::Vector{Float64})
     # Parameters
     start_genome = "AAAA"
     optimal_genome = "BBBB"
+    genome_length = length(optimal_genome)
 
     pat_type = jOpqua.newPathogenType(
         "pat_type",
-        num_loci=4,
+        num_loci=genome_length,
         possible_alleles="AB",
-        # mean_effective_inoculum=1.0,
-        # mean_mutations_per_replication=0.00003,
-        contactSpecificCoefficient=s::String -> 1.0 + (0.1 * (4.0 - hamming(s, optimal_genome)) / 4.0),
-        receiveContactHostwideCoefficient=s::String -> 0.0,
+        contactSpecificCoefficient=s::String -> 1.0 + (0.1 * (genome_length - hamming(s, optimal_genome)) / genome_length),
+        receiveContactHostwideCoefficient=s::String -> 0.0, # makes infected hosts immune to superinfection
     )
 
     res_type = jOpqua.newResponseType(
         "res_type",
         reactivityCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 1.0 : 0.0,
-        transmissionEfficiencyInteractionSpecificCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 0.0 : 1.0,
-        clearanceInteractionSpecificCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 1.0e3 : 1.0,
+        # transmissionEfficiencyInteractionSpecificCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 0.0 : 1.0,
+        clearanceInteractionSpecificCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 1.1e0 : 1.0,
         # responseAcquisitionInteractionSpecificCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 0.0 : 1.0,
     )
 
@@ -41,13 +40,16 @@ function run(seed::Int64, t_vec::Vector{Float64})
 
     pop_type = jOpqua.newPopulationType(
         "pop_type",
-        clearance_coefficient=1.0e-3,
+        clearance_coefficient=1.0,
         contact_coefficient=1.05,
-        response_acquisition_coefficient=0.5,
-        response_loss_coefficient=0.0,
+        # response_acquisition_coefficient=0.5,
+        response_acquisition_upon_clearance_coefficient=0.1,
+        response_loss_coefficient=0.01,
         receive_contact_coefficient=1.0,
-        mutations_upon_infection_coefficient=0.00003,
+        mutations_per_generation_coefficient=0.0005,
         inoculum_coefficient=1.0,
+        # death_coefficient=0.001,
+        # birth_coefficient=0.001,
         pathogenFractions=jOpqua.pathogenFractionsProportionalFitness,
         response_types=Dict{String,jOpqua.ResponseType}([(res_type.id => res_type)]),
         developResponses=(
@@ -55,13 +57,13 @@ function run(seed::Int64, t_vec::Vector{Float64})
             existing_responses::Dict{Tuple{String,String,String,String},jOpqua.Response},
             response_types::Dict{String,jOpqua.ResponseType},
             birth_time::Float64
-        ) -> jOpqua.deNovoResponse(
+        ) -> [jOpqua.deNovoResponse(
             pathogen, host, existing_responses, response_types, birth_time;
             response_type_id="res_type"
-        ),
+        )],
     )
 
-    num_hosts = 50000
+    num_hosts = 10000
     num_infected = Int(num_hosts * 0.05)
     host_genome = ""
 
@@ -88,17 +90,16 @@ function run(seed::Int64, t_vec::Vector{Float64})
     his_dat = jOpqua.saveHistory(output, "examples/acquired_immunity/history_acquired_immunity.csv")
     composition_data = jOpqua.saveComposition(
         his_dat, "examples/acquired_immunity/composition_acquired_immunity.csv",
-        num_top_sequences=7, track_specific_sequences=["AAAA", "BBBB"]
+        num_top_sequences=7, #track_specific_sequences=["AAAA", "BBBB"]
     )
     jOpqua.plotComposition(
         composition_data, "examples/acquired_immunity/composition_acquired_immunity.png",
-        # normalized=true, ylabel="Fraction",
-        normalized=false, ylabel="Number",
+        normalized=true, ylabel="Fraction",
     )
 
     composition_data = jOpqua.saveComposition(
         his_dat, "examples/acquired_immunity/composition_acquired_immunity_responses.csv",
-        num_top_sequences=7, track_specific_sequences=["AAAA", "BBBB"],
+        num_top_sequences=7, #track_specific_sequences=["AAAA", "BBBB"],
         type_of_composition="Response_imprinted_sequence"
     )
     jOpqua.plotComposition(
@@ -107,11 +108,16 @@ function run(seed::Int64, t_vec::Vector{Float64})
         normalized=false, ylabel="Number",
     )
 
-    nwks = jOpqua.saveNewick(output, "examples/acquired_immunity/pathogen_newick_acquired_immunity.nwk")
-    for nwk in nwks
-        jOpqua.plotPhylogeny(nwk, "examples/acquired_immunity/pathogen_newick_acquired_immunity.png")
-    end
+    # nwks = jOpqua.saveNewick(output, "examples/acquired_immunity/pathogen_newick_acquired_immunity.nwk", branch_length="Time", info_separator=" ")
+    # for nwk in nwks
+    #     jOpqua.plotPhylogeny(nwk, "examples/acquired_immunity/pathogen_newick_acquired_immunity.png")
+    # end
 end
 
 run(1, collect(0.0:2.0:4.0)) # compile
-@time run(0, collect(0.0:0.20:100.0))
+@time run(10, collect(0.0:2.0:1500.0))
+# Total events: 3342817
+#   7.751170 seconds (82.72 M allocations: 14.894 GiB, 20.86% gc time, 7.99% compilation time: <1% of which was recompilation)
+# 25 May 2026 Julia 1.12.6 Apple M3 Max 128 GB RAM
+
+# @profview run(2, collect(0.0:2.0:1500.0))